+++ samples/sandbox/src/main/kotlin/com/bumble/appyx/sandbox/client/cardsexample/CardsExampleNode.kt
@@ -4,7 +4,6 @@ import android.os.Parcelable
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.coroutineScope